64bit_to_32bit
64bit_to_32bit

+++ src/pandemia/components/health_model/default_health_model.py
@@ -133,13 +133,13 @@ class DefaultHealthModel(HealthModel):
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
                       self.number_of_strains,
-                      self.max_preset_length_immunity), dtype=np.int64)
+                      self.max_preset_length_immunity), dtype=np.int32)
 
         self.preset_rho_immunity_failure_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         self.preset_sigma_immunity_failure_values =\
             np.zeros((self.number_of_presets,
@@ -153,13 +153,13 @@ class DefaultHealthModel(HealthModel):
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
                       self.number_of_strains,
-                      self.max_preset_length_immunity), dtype=np.int64)
+                      self.max_preset_length_immunity), dtype=np.int32)
 
         self.preset_sigma_immunity_failure_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         self.preset_infectiousness_values =\
             np.zeros((self.number_of_presets,
@@ -171,12 +171,12 @@ class DefaultHealthModel(HealthModel):
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.int64)
+                      self.max_preset_length_health), dtype=np.int32)
 
         self.preset_infectiousness_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         self.preset_disease_values =\
             np.zeros((self.number_of_presets,
@@ -188,29 +188,29 @@ class DefaultHealthModel(HealthModel):
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.int64)
+                      self.max_preset_length_health), dtype=np.int32)
 
         self.preset_disease_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         self.preset_strain_values =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.int64)
+                      self.max_preset_length_health), dtype=np.int32)
 
         self.preset_strain_partitions =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
                       self.number_of_strains,
-                      self.max_preset_length_health), dtype=np.int64)
+                      self.max_preset_length_health), dtype=np.int32)
 
         self.preset_strain_lengths =\
             np.zeros((self.number_of_presets,
                       self.number_of_rho_immunity_outcomes,
-                      self.number_of_strains), dtype=np.int64)
+                      self.number_of_strains), dtype=np.int32)
 
         # Create presets, thereby filling the above arrays, and validate them
         self._get_presets(self.clock.ticks_in_day)
@@ -259,11 +259,11 @@ class DefaultHealthModel(HealthModel):
         immunity_length                 = self.immunity_length
         max_preset_length_health        = self.max_preset_length_health
 
-        vector_region.health_age_group                 = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.health_age_group                 = np.zeros((number_of_agents), dtype=np.int32)
         vector_region.location_transmission_multiplier = np.ones((number_of_locations), dtype=np.float64)
-        vector_region.infection_event                  = np.full((number_of_agents), -1, dtype=np.int64)
-        vector_region.presets                          = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.requesting_immunity_update       = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.infection_event                  = np.full((number_of_agents), -1, dtype=np.int32)
+        vector_region.presets                          = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.requesting_immunity_update       = np.zeros((number_of_agents), dtype=np.int32)
 
         vector_region.current_rho_immunity_failure   = np.ones((number_of_agents,
                                                                 number_of_strains,
@@ -275,7 +275,7 @@ class DefaultHealthModel(HealthModel):
 
         vector_region.current_infectiousness = np.zeros((number_of_agents), dtype=np.float64)
         vector_region.current_disease        = np.zeros((number_of_agents), dtype=np.float64)
-        vector_region.current_strain         = np.full((number_of_agents), -1, dtype=np.int64)
+        vector_region.current_strain         = np.full((number_of_agents), -1, dtype=np.int32)
 
         vector_region.rho_immunity_failure_values   = np.ones((number_of_agents,
                                                                number_of_strains,
@@ -290,21 +290,21 @@ class DefaultHealthModel(HealthModel):
         vector_region.infectiousness_values     = np.zeros((number_of_agents,
                                                             max_preset_length_health), dtype=np.float64)
         vector_region.infectiousness_partitions = np.zeros((number_of_agents,
-                                                            max_preset_length_health), dtype=np.int64)
-        vector_region.infectiousness_lengths    = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.infectiousness_indexes    = np.ones((number_of_agents), dtype=np.int64)
+                                                            max_preset_length_health), dtype=np.int32)
+        vector_region.infectiousness_lengths    = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.infectiousness_indexes    = np.ones((number_of_agents), dtype=np.int32)
         vector_region.disease_values            = np.zeros((number_of_agents,
                                                             max_preset_length_health), dtype=np.float64)
         vector_region.disease_partitions        = np.zeros((number_of_agents,
-                                                            max_preset_length_health), dtype=np.int64)
-        vector_region.disease_lengths           = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.disease_indexes           = np.ones((number_of_agents), dtype=np.int64)
+                                                            max_preset_length_health), dtype=np.int32)
+        vector_region.disease_lengths           = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.disease_indexes           = np.ones((number_of_agents), dtype=np.int32)
         vector_region.strain_values             = np.full((number_of_agents,
-                                                           max_preset_length_health), -1, dtype=np.int64)
+                                                           max_preset_length_health), -1, dtype=np.int32)
         vector_region.strain_partitions         = np.zeros((number_of_agents,
-                                                            max_preset_length_health), dtype=np.int64)
-        vector_region.strain_lengths            = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.strain_indexes            = np.ones((number_of_agents), dtype=np.int64)
+                                                            max_preset_length_health), dtype=np.int32)
+        vector_region.strain_lengths            = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.strain_indexes            = np.ones((number_of_agents), dtype=np.int32)
 
         self._get_age_mixing_matrix(vector_region)
 
@@ -755,7 +755,7 @@ class DefaultHealthModel(HealthModel):
                 else:
                     vector_region.subpopulation_mixing_matrix = np.ones((1, 1), dtype=np.float64)
                     vector_region.number_of_subpopulations = 1
-            vector_region.age_group = np.zeros((vector_region.number_of_agents), dtype=np.int64)
+            vector_region.age_group = np.zeros((vector_region.number_of_agents), dtype=np.int32)
             for n in range(vector_region.number_of_agents):
                 age_group = min(vector_region.age[n] // self.age_group_interval,
                                 vector_region.number_of_subpopulations - 1)
@@ -763,7 +763,7 @@ class DefaultHealthModel(HealthModel):
         else:
             vector_region.subpopulation_mixing_matrix = np.ones((1, 1), dtype=np.float64)
             vector_region.subpopulation_index =\
-                np.zeros((vector_region.number_of_agents), dtype=np.int64)
+                np.zeros((vector_region.number_of_agents), dtype=np.int32)
             vector_region.number_of_subpopulations = 1
 
     def _generate_presets_and_weights(self, sir_beta, sir_gamma_inverse,
