+++ src/pandemia/components/hospitalization_and_death_model/default_hospitalization_and_death_model.py
@@ -53,10 +53,10 @@ class DefaultHospitalizationAndDeathModel(HospitalizationAndDeathModel):
 
         vector_region.number_of_hospitals = len(hospitals)
         vector_region.number_of_cemeteries = len(cemeteries)
-        vector_region.hospitals = np.array(hospitals).astype(np.int64)
-        vector_region.cemeteries = np.array(cemeteries).astype(np.int64)
-        vector_region.in_hospital = np.zeros((number_of_agents), dtype=np.int64)
-        vector_region.in_cemetery = np.zeros((number_of_agents), dtype=np.int64)
+        vector_region.hospitals = np.array(hospitals).astype(np.int32)
+        vector_region.cemeteries = np.array(cemeteries).astype(np.int32)
+        vector_region.in_hospital = np.zeros((number_of_agents), dtype=np.int32)
+        vector_region.in_cemetery = np.zeros((number_of_agents), dtype=np.int32)
 
     def initial_conditions(self, vector_region):
         """Establishes initial conditions for hospitalization and death."""
